+++ choco-solver/src/main/java/org/chocosolver/solver/variables/impl/IntervalIntVarImpl.java
@@ -325,12 +325,6 @@ public final class IntervalIntVarImpl extends AbstractVariable implements IntVar
 
 
     @Override
-    public void wipeOut(ICause cause) throws ContradictionException {
-        assert cause != null;
-        removeInterval(this.getLB(), this.getUB(), cause);
-    }
-
-    @Override
     public boolean isInstantiated() {
         return SIZE.get() == 1;
     }